Move contexts to @epam/uui-docs
