+++ app/src/common/docs/componentEditor/utils.ts
@@ -1,5 +1,5 @@
 import { IPropSamplesCreationContext, PropDoc, PropExampleObject, TSkin } from '@epam/uui-docs';
-import { TTheme } from '../BaseDocsBlock';
+import { TTheme, TUUITheme } from '../docsConstants';
 
 export function isPropValueEmpty(propValue: any): boolean {
     return propValue === undefined;
