+++ app/src/docs/_props/uui/docs/index.ts
@@ -1,4 +1,3 @@
 export * from './common';
-export * from './contexts';
 export * from './../components/pickers/common';
 export * from './../components/iconPicker/IconPicker';
