+++ uui-docs/src/contexts/ResizableContext.module.scss
