+++ uui-docs/src/contexts/ResizableContext.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { DemoComponentProps } from '@epam/uui-docs';
+import { DemoComponentProps } from '../types';
 import { Slider } from '@epam/uui';
 import css from './ResizableContext.module.scss';
