+++ uui-docs/src/index.ts
@@ -6,3 +6,4 @@ export * from './demoData';
 export * from './models';
 export * from './commonDocs';
 export * from './editors';
+export * from './contexts/index';
